crud nguyên liệu trong tử lạnh

+++ src/main/java/org/example/repository/indoor/UserFridgeRepository.java
@@ -3,9 +3,10 @@ package org.example.repository.indoor;
 import org.example.entity.indoor.UserFridge;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 @Repository
-public interface UserFridgeRepository extends JpaRepository<UserFridge, Long> {
-    List<UserFridge> findByUserId(Long userId);
+public interface UserFridgeRepository extends JpaRepository<UserFridge, Integer> {
+    List<UserFridge> findByUserId(Integer userId);
 }
